+++ dspy_implementation/dc_module/__init__.py
@@ -1,0 +1,10 @@
+"""
+Dynamic Cheatsheet Module for MMESGBench
+Integrates DC test-time learning with our RAG pipeline
+"""
+
+from .dc_wrapper import DCWrapper
+from .dc_rag_module import DCRAGModule
+
+__all__ = ['DCWrapper', 'DCRAGModule']
+